refactor: update authentication service imports and error handling
- Replaced references to the old authService module with the new auth module in LockScreen, LoginScreen, and SignUpScreen.
- Simplified password reset logic by directly using the new authService methods.
- Enhanced error handling to provide clearer messages based on API error responses.
